requirements.txt
haystack-ai[all]
chromadb
chroma-haystack
Ollama-haystack
pytest